refactor(chains): renamed pkg/chain to pkg/chains
