+++ pkg/chains/evm_adapter_test.go
@@ -1,4 +1,4 @@
-package chain
+package chains
 
 import (
 	"context"
